+++ lib/fpath_sexp0/src/fpath_sexp0.ml
@@ -13,4 +13,3 @@ end
 module Absolute_path = Path.Absolute_path
 module Relative_path = Path.Relative_path
 module Fsegment = Fsegment
-module Fpart = Fsegment