+++ lib/fpath_sexp0/test/test__fpart.mli
@@ -1,5 +1,0 @@
-(*_********************************************************************************)
-(*_  fpath-base: Extending [Fpath] to use alongside [Sexplib0] and/or [Base]      *)
-(*_  SPDX-FileCopyrightText: 2023-2025 Mathieu Barbin <mathieu.barbin@gmail.com>  *)
-(*_  SPDX-License-Identifier: MIT                                                 *)
-(*_********************************************************************************)